feat: add routed module placeholders
