+++ src/App.css
@@ -118,6 +118,7 @@
   transition: background 0.2s ease, border-color 0.2s ease, color 0.2s ease;
   cursor: pointer;
   position: relative;
+  text-decoration: none;
 }
 
 .nav-button:hover,
@@ -220,6 +221,7 @@
   cursor: pointer;
   transition: background 0.2s ease, border-color 0.2s ease, color 0.2s ease;
   color: var(--text-secondary);
+  position: relative;
 }
 
 .icon-button:hover,
@@ -227,6 +229,17 @@
   border-color: var(--border);
   background: var(--surface-muted);
   color: var(--ink-500);
+}
+
+.icon-button-indicator {
+  position: absolute;
+  width: 9px;
+  height: 9px;
+  border-radius: 999px;
+  background: var(--warning);
+  top: 10px;
+  right: 10px;
+  box-shadow: 0 0 0 4px rgba(246, 167, 35, 0.2);
 }
 
 .card {
@@ -635,6 +648,48 @@
 .knowledge-tree {
   display: grid;
   gap: 20px;
+}
+
+.under-construction {
+  min-height: 420px;
+  display: flex;
+  flex-direction: column;
+}
+
+.empty-state {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  gap: 20px;
+  padding: 24px;
+  background: linear-gradient(135deg, rgba(255, 231, 204, 0.35), rgba(247, 242, 231, 0.6));
+  border-radius: var(--radius-card);
+  border: 1px dashed rgba(50, 75, 92, 0.18);
+  color: var(--text-secondary);
+}
+
+.empty-state p {
+  margin: 0;
+  font-size: 17px;
+  color: var(--ink-500);
+}
+
+.empty-state-list {
+  margin: 0;
+  padding-left: 18px;
+  display: grid;
+  gap: 8px;
+  color: var(--ink-400);
+}
+
+.empty-state-footer {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  color: var(--ink-300);
+  font-size: 14px;
 }
 
 .branch {
